requirements.txt
telethon==1.42.0
httpx==0.28.1
python-dotenv==1.1.1
aiosqlite==0.21.0
orjson==3.11.3
fastapi==0.115.6
uvicorn[standard]==0.32.1
